tools/embed/requirements.txt
haystack-ai
ollama-haystack
elasticsearch-haystack
python-dotenv
huggingface_hub
